Moved classes to sub-packages
